summarizer-github/requirements.txt
fastapi
uvicorn
python-dotenv
pypdf2
